// libs/feature/metrics/src/lib/generated/prisma-client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck
/**
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */
export const MetricsRole = {
  Admin: 'Admin',
  User: 'User',
} as const;

export type MetricsRole = (typeof MetricsRole)[keyof typeof MetricsRole];